[AN] refactor(#495): api 서비스 관련 패키지 수정
