+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/feed/StudyApplyViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.takseha.data.dto.feed.MessageRequest
 import com.takseha.data.dto.mystudy.StudyInfoResponse
-import com.takseha.data.repository.member.GitudyMemberRepository
-import com.takseha.data.repository.study.GitudyStudyRepository
+import com.takseha.data.repository.gitudy.GitudyMemberRepository
+import com.takseha.data.repository.gitudy.GitudyStudyRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
