+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/mystudy/AddTodoViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.takseha.data.dto.mystudy.MakeTodoRequest
-import com.takseha.data.repository.study.GitudyStudyRepository
+import com.takseha.data.repository.gitudy.GitudyStudyRepository
 import kotlinx.coroutines.launch
 
 class AddTodoViewModel: ViewModel() {
