+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/mystudy/CommitConventionViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.takseha.data.dto.mystudy.SetConventionRequest
-import com.takseha.data.repository.study.GitudyStudyRepository
+import com.takseha.data.repository.gitudy.GitudyStudyRepository
 import kotlinx.coroutines.launch
 
 // 추후 삭제 예정
