+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/viewmodel/mystudy/StudyCommentBoardViewModel.kt
@@ -4,16 +4,9 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.takseha.data.dto.mystudy.StudyComment
-import com.takseha.data.dto.mystudy.StudyConvention
-import com.takseha.data.dto.mystudy.StudyInfoResponse
-import com.takseha.data.dto.mystudy.StudyMember
-import com.takseha.data.dto.mystudy.Todo
-import com.takseha.data.dto.mystudy.TodoProgressResponse
-import com.takseha.data.repository.member.GitudyMemberRepository
-import com.takseha.data.repository.study.GitudyStudyRepository
+import com.takseha.data.repository.gitudy.GitudyStudyRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class StudyCommentBoardViewModel : ViewModel() {
